Refaktorering av login-knapp. Hindrer overstyring av styling fra andre apper.
