--- src/komponenter/header/header-regular/common/logg-inn/LoggInn.less
@@ -1,7 +1,0 @@
-@import '../../../../../styling-variabler.less';
-
-.login-knapp {
-    text-transform: none;
-    letter-spacing: initial;
-    font: initial;
-}
