+++ src/komponenter/header/header-regular/desktop/DesktopMenylinje.less
@@ -76,7 +76,7 @@
         max-height: @maxheight-menuitem;
     }
 
-    .login-knapp {
+    .login-knapp-container {
         order: 7;
         max-height: @maxheight-menuitem;
         margin-right: @margins-sides-desktop;
