+++ src/komponenter/header/header-regular/desktop/DesktopMenylinje.tsx
@@ -5,24 +5,25 @@ import { SokDropdown } from './sok/SokDropdown';
 import { HovedmenyDesktop } from './hovedmeny/HovedmenyDesktop';
 import { MinsideMeny } from './minside-meny/MinsideMeny';
 import { VarslerDropdown } from './varsler/VarslerDropdown';
-import { LoggInnKnappDesktop } from './logg-inn/LoggInnKnappDesktop';
 import { useKbNavMain } from 'utils/keyboard-navigation/useKbNavMain';
 import { MenuValue } from 'utils/meny-storage-utils';
 import { useSelector } from 'react-redux';
 import { AppState } from 'store/reducers';
+import LoggInnKnapp from 'komponenter/header/header-regular/common/logg-inn-knapp/LoggInnKnapp';
 import './DesktopMenylinje.less';
 
 export const desktopHeaderLogoId = 'desktop-header-logo-id';
+export const desktopLoginKnappId = 'desktop-login-knapp';
 
 const DesktopMenylinje = () => {
     const cls = BEMHelper('desktopmeny');
     const kbNavMainState = useKbNavMain();
     const arbeidsflate = useSelector(
-        (state: AppState) => state.arbeidsflate.status
+        (state: AppState) => state.arbeidsflate.status,
     );
 
     const innlogga = useSelector(
-        (state: AppState) => state.innloggingsstatus.data.authenticated
+        (state: AppState) => state.innloggingsstatus.data.authenticated,
     );
 
     const visVarslerDropdown =
@@ -44,7 +45,7 @@ const DesktopMenylinje = () => {
                     <VarslerDropdown kbNavMainState={kbNavMainState} />
                 )}
                 <MinsideMeny kbNavMainState={kbNavMainState} />
-                <LoggInnKnappDesktop />
+                <LoggInnKnapp id={desktopLoginKnappId} />
             </div>
         </nav>
     );
