--- src/komponenter/header/header-regular/desktop/logg-inn/LoggInnKnappDesktop.less
@@ -1,5 +1,0 @@
-@import '../../../../../styling-variabler.less';
-
-.desktop-login-knapp {
-    padding: 0.625rem 1rem;
-}
