+++ src/komponenter/header/header-regular/mobil/MobilMenylinje.less
@@ -58,7 +58,7 @@
         }
     }
 
-    .login-knapp {
+    .login-knapp-container {
         order: 2;
         margin-right: 3px;
     }
