--- src/komponenter/header/header-regular/mobil/logg-inn/LoggInnKnappMobil.less
@@ -1,10 +1,0 @@
-@import '../../../../../styling-variabler.less';
-
-.mobil-login-knapp {
-    padding: 0.5rem 1rem;
-    margin-left: 3px;
-
-    @media (max-width: @minimumMobilWidth) {
-        padding: 0.5rem 0.5rem;
-    }
-}
