+++ src/styling-variabler.less
@@ -2,6 +2,7 @@
 @import (reference) '~nav-frontend-lenker-style/src/lenker-style';
 @import (reference) '~nav-frontend-core/less/_variabler';
 @import (reference) '~nav-frontend-typografi-style/src/index';
+@import (reference) '~nav-frontend-knapper-style/src/index';
 
 //desktop breakpoint width
 @desktopLimitWidth: 1440px;
